requirements.txt
asgiref==3.5.2
autopep8==2.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django>=4.0.0
django-cors-headers==3.13.0
django-import-export==3.0.1
djangorestframework==3.14.0
djangorestframework-camel-case==1.3.0
drf-yasg==1.21.4
et-xmlfile==1.1.0
flake8==5.0.4
flake8-django==1.1.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
openpyxl==3.0.10
packaging==21.3
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pytz==2022.6
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
tablib==3.2.1
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.12
xlrd==2.0.1
xlwt==1.3.0
djangorestframework-simplejwt>=5.2.0